@@ -1,7 +1,4 @@
-﻿using OpenTK.Audio.OpenAL;
-
-
-namespace demo.common.audio.impl.al;
+﻿namespace demo.common.audio.impl.al;
 
 public partial class AlAudioManager {
   public IAudioSource<short> CreateAudioSource() => new AlAudioSource(this);
