@@ -1,16 +1,11 @@
-﻿using System.Numerics;
+﻿namespace libsm64sharp;
 
-
-namespace libsm64sharp;
-
-public interface IReadOnlySm64Vector2<TNumber>
-    where TNumber : INumber<TNumber> {
-  TNumber X { get; }
-  TNumber Y { get; }
+public interface IReadOnlySm64Vector2f {
+  float X { get; }
+  float Y { get; }
 }
 
-public interface ISm64Vector2<TNumber> : IReadOnlySm64Vector2<TNumber>
-    where TNumber : INumber<TNumber> {
-  new TNumber X { get; set; }
-  new TNumber Y { get; set; }
+public interface ISm64Vector2f : IReadOnlySm64Vector2f {
+  new float X { get; set; }
+  new float Y { get; set; }
 }
