@@ -1,13 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Windows.Forms;
-
-namespace Quad64.src.Forms;
+﻿namespace Quad64.src.Forms;
 
 public partial class EditWarp : Form
 {
