@@ -1,8 +1,5 @@
 ﻿using Quad64.src;
-using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Windows.Forms;
+
 
 namespace Quad64;
 
