@@ -1,13 +1,7 @@
-﻿using Quad64.src.JSON;
-using Quad64.src.LevelInfo;
+﻿using Quad64.src.LevelInfo;
 using Quad64.src.Scripts;
 
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Drawing;
 using System.Runtime.InteropServices;
-using System.Windows.Forms;
 
 
 namespace Quad64.src.Forms;
