@@ -1,10 +1,5 @@
-﻿using Quad64.src.LevelInfo;
-using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Windows.Forms;
+﻿using System.Text;
+
 
 namespace Quad64.src.Forms;
 
