@@ -1,13 +1,7 @@
 ﻿using Quad64.src.JSON;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
+
 using System.Runtime.InteropServices;
-using System.Text;
-using System.Windows.Forms;
+
 
 namespace Quad64.src.Forms;
 
