@@ -1,13 +1,5 @@
 ﻿using Quad64.src;
-using Quad64.src.LevelInfo;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Windows.Forms;
+
 
 namespace Quad64;
 
