@@ -1,11 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Windows.Forms;
-
-namespace Quad64.src.Forms.TextureEditorComponents;
+﻿namespace Quad64.src.Forms.TextureEditorComponents;
 
 class RadioButtonWithInfo : RadioButton
 {
